completed

+++ app/controllers/tweets_controller.rb
@@ -1,50 +1,45 @@
 class TweetsController < ApplicationController
+  def index
+    @tweets = Tweet.all.order(created_at: :desc)
+    render 'tweets/index'
+  end
+
   def create
     token = cookies.signed[:twitter_session_token]
     session = Session.find_by(token: token)
-    
-    if session
-      user = session.user
-      @tweet = user.tweet.new(tweet_params)
+    user = session.user
+    @tweet = user.tweets.new(tweet_params)
 
-      if tweet.save
-        render 'tweets/create'
-      else
-        render json: { success: false}
-      end
-    else
-      render json: { success: false}
-    end    
+    render 'tweets/create' if @tweet.save
   end
 
   def destroy
-    @tweet = Tweet.find_by(id: params[:id])
     token = cookies.signed[:twitter_session_token]
     session = Session.find_by(token: token)
 
-    if session
-      user = session.user
-      if user == tweet.user
-        tweet.destroy
-      end
+    return render json: { success: false } unless session
+
+    user = session.user
+    tweet = Tweet.find_by(id: params[:id])
+
+    if tweet && (tweet.user == user) && tweet.destroy
+      render json: {
+        success: true
+      }
+    else
+      render json: {
+        success: false
+      }
     end
-
-  end
-
-  def index
-    @tweets = Tweet.all
   end
 
   def index_by_user
-    token = cookies.signed[:twitter_session_token]
-    session = Session.find_by(token: token)
+    user = User.find_by(username: params[:username])
 
-    if session
-      @tweets = session.user.tasks
-    else
-      render json: { tasks: [] }
+    if user
+      @tweets = user.tweets
+      render 'tweets/index'
     end
-
   end
 
   private
@@ -52,5 +47,4 @@ class TweetsController < ApplicationController
   def tweet_params
     params.require(:tweet).permit(:message)
   end
-
 end
